Fixes1 in directory

+++ Assets/Scripts/TrapDoor.cs
@@ -34,25 +34,25 @@ public class TrapDoor : MonoBehaviour
         {
             if(SceneManager.GetActiveScene().name == "Level1")
             {
-                PlayerPrefs.SetInt("PlayerLivesFromLevel1", GameManager.instance().GetPlayerLives());
+                PlayerPrefs.SetFloat("PlayerLivesFromLevel1", GameManager.instance().GetPlayerLives());
                 SceneManager.LoadScene("Level2");
                 GameManager.instance().UpdateSaveSession();
             }
             else if(SceneManager.GetActiveScene().name == "Level2") 
             {
-                PlayerPrefs.SetInt("PlayerLivesFromLevel1", GameManager.instance().GetPlayerLives());
+                PlayerPrefs.SetFloat("PlayerLivesFromLevel1", GameManager.instance().GetPlayerLives());
                 SceneManager.LoadScene("Level3");
                 GameManager.instance().UpdateSaveSession();
             }
             else if(SceneManager.GetActiveScene().name == "Level3") 
             {
-                PlayerPrefs.SetInt("PlayerLivesFromLevel1", GameManager.instance().GetPlayerLives());
+                PlayerPrefs.SetFloat("PlayerLivesFromLevel1", GameManager.instance().GetPlayerLives());
                 SceneManager.LoadScene("Level4");
                 GameManager.instance().UpdateSaveSession();
             }
             else
             {
-                PlayerPrefs.SetInt("PlayerLivesFromLevel1", 3);
+                PlayerPrefs.SetFloat("PlayerLivesFromLevel1", 3f);
                 Time.timeScale = 0f;
                 SceneManager.LoadScene("WinTheGame");
             }
